requirements.txt
python-telegram-bot==20.9.1
Flask